+++ OneBlink.SDK/models/WorkspaceMapping.cs
@@ -1,0 +1,14 @@
+namespace OneBlink.SDK.Model
+{
+    public class WorkspaceMapping
+    {
+        public long sourceWorkspaceId
+        {
+            get; set;
+        }
+        public long targetWorkspaceId
+        {
+            get; set;
+        }
+    }
+}